Move AccessMode types to shared types
Instead of gconsent-specific ones
