+++ src/index.ts
@@ -24,7 +24,6 @@ export type {
   AccessBaseOptions,
   AccessCredentialType,
   AccessGrantContext,
-  AccessModes,
   AccessRequest,
   ApproveAccessRequestOverrides,
   IssueAccessRequestParameters,
@@ -33,6 +32,8 @@ export type {
 
 export type { FetchOptions } from "./type/FetchOptions";
 export type { RedirectOptions } from "./type/RedirectOptions";
+export type { AccessModes } from "./type/AccessModes";
+
 
 export {
   getAccessApiEndpoint,
